refactor: `OAuth2Response` 위치 이동
